+++ glances_api/exceptions.py
@@ -6,14 +6,18 @@ class GlancesApiError(Exception):
 
     pass
 
+
 class GlancesApiConnectionError(GlancesApiError):
     """When a connection error is encountered."""
 
     pass
+
+
 class GlancesApiAuthorizationError(GlancesApiError):
     """When a connection error is encountered."""
 
     pass
+
 
 class GlancesApiNoDataAvailable(GlancesApiError):
     """When no data is available."""